Check if file exists in okbuck clean

+++ buildSrc/src/main/java/com/uber/okbuck/core/task/OkBuckCleanTask.java
@@ -82,8 +82,11 @@ public class OkBuckCleanTask extends DefaultTask {
     }
 
     private static void deleteQuietly(Path p) {
+        File f = p.toFile();
+        if (!f.exists()) {
+            return;
+        }
         try {
-            File f = p.toFile();
             if (f.isDirectory()) {
                 FileUtils.deleteDirectory(f);
             } else {
